UberConf wrapper

OO API for reading UberConf new Constructor get_all_pdo_name Get all known program degree offering names get_pdo Get full UberConf tree for some program degree offering. Does not compute any derived values. get_all_value Get UberConf value across all program degree offerings. Computes derived values as in get_pdo_value. get_pdo_value Get UberConf value for a program degree offering and specific key. Computes derived values for some known keys.

Arguments

...
Passed to get_uber_conf

pdo
Program degree offering like unc-mba

key
UberConf key like salesforce_org

path
Path on disk, passed through path.expand

global
Read UberConf data into an environment, so there's a global copy. If FALSE, read data into a list, i.e., a copy that only exists within one instance of this class.

reload
Force a reload of a global copy

Format

R6Class object.

Examples

not_run({
  u <- UberConf$new() # Check out from GitHub
  u <- UberConf$new(path="~/git/uber-conf/data") # Read from disk
  u$get_pdo_value("unc-mba", "salesforce_org") # Key defined in UberConf
  u$get_pdo_value("unc-mba", "segment_site") # Key yields derived value
})