UberConf wrapper

OO API for modifying and writing UberConf new Constructor get_all_pdo_name Get all known program degree offering names get_pdo Get full UberConf tree for some program degree offering. Does not compute any derived values. set_all_key_value Sets a constant value for all program degree offerings. set_pdo_key_value Sets a value for one program degree offering. write_json_all Writes JSON files for all program degree offerings. write_json_pdo Writes JSON file for one program degree offering.

Arguments

uber
an object of class UberConf

pdo
Program degree offering like unc-mba

key_value
list like list(sis=list(sis_bucket="au-mba"))

path
Path on disk, passed through path.expand

keep.null
If false, removes entries that are set to NULL

Format

R6Class object.

Examples

not_run({
  uw <- UberConfWriter$new(UberConf$new())
  n <- list(sis = list(sis_bucket = NULL, university_id = NULL))
  uw$set_all_key_value(n)
  uw$write_json_all("mydir")
})