send

Execute arbitrary sql, expecting no response. Str_form substitution params supported by default.

Execute arbitrary sql, expecting no response. Str_form substitution params supported by default.

Arguments

database
Name of database for which the connection should created. Must correspond to a creds.json file key.

cred_location
Path to a json credentials file with keys corresponding to database names. Each database entry should have the following fields: db_name, user, password, host, port Defaults to creds.json in user root directory.

database_type
One of postgres, mysql or ms_sql (for fdw connections). Defaults to postgres

init_sql
A sql statement to execute on connection; optional.

table
Name of table for which columns should be returned.

match_string
String to search for in columns. Partial matching applied. Optional.

Details

Delivering dead simple database connections since June 7th 2016.

Fields

connection

A dplyr src connection object used to execute queries.

con

A DBI compliant connection, extracted from dplyr_src.

Methods

get_cols(table, match_string = NA)

Return a list of columns in the specified table. Can optionally apply a regex match to find a specific column.

query(statement, ...)

Execute query using initialized database connection. Str_form templating is supported by default.

send(statement, ...)

Send a statement using initialized database connection. Str_form templating is supported by default.