export_jobs_to_jenkins

Exports the pre-populated xml files in the jenkins directory to the server.

export_jobs_to_jenkins(user = yaml::yaml.load_file("jenkins/deploy.yaml")$user_email,
  api_token = Sys.getenv("jenkins_api_token"),
  project = yaml::yaml.load_file("jenkins/deploy.yaml")$project_name,
  host = "https://corp-bi-jenkins-prod.2u.com")

Arguments

user
Jenkins user email address used to authenticate. Defaults to project user email found in jenkins deploy.yaml file.

api_token
The api token associated with the jenkins user being used. By default, mmkit will look for a system enviroment var called jenkins_api_token.

host
Hostname of the jenkins server where jobs should be created. Defaults to https://corp-bi-jenkins-prod.2u.com. Don't change this.

project_name
Project name - Used as prefix for jobs being created. By default, mmkit will try to read this information from the jenkins/deploy.yaml file.